requirements.txt
streamlit>=1.30.0
openai>=1.0.0
fpdf2>=2.8.0
python-dotenv>=1.0.0
